+++ packages/localizations/src/ko-KR.ts
@@ -52,6 +52,9 @@ export const koKR: LocalizationResource = {
       formHint: '이 시크릿 키를 삭제할까요?',
       formTitle: '"{{apiKeyName}}" 시크릿 키를 폐기할까요?',
     },
+    tableHeader__name: '이름',
+    tableHeader__lastUsed: '마지막 사용',
+    tableHeader__actions: '동작',
   },
   backButton: '돌아가기',
   badge__activePlan: '활성',
@@ -181,6 +184,11 @@ export const koKR: LocalizationResource = {
     viewFeatures: '기능 보기',
     viewPayment: '결제 보기',
     year: '년',
+  },
+  configureSSO: {
+    navbar: {
+      title: '싱글 사인온(SSO) 구성',
+    },
   },
   createOrganization: {
     formButtonSubmit: '조직 만들기',